Fix the C++ OpenGL compositor on Linux
We used to tell OpenGL that the length of our texture coordinates vector is 3.
This is obviously wrong, since texture coordinates are 2-dimensional.
Native macOS OpenGL driver and Google ANGLE were both able to figure out from
the shader code that this is the case and correct it (that's what I
think at least), however Linux drivers didn't. Now the code is correct
and works!