Make a new `RawVideo` struct and rename the old one to `ElixirRawVideo`.
Using rustler structs (such as atoms) outside of the rust-elixir
interface layer made us unable to use rust's unit tests.

This patch adds a new `RawVideo` struct, which is similiar to the old
one, but has a rust enum instead of a `rustler::Atom` for
`pixel_format`.

Since I was already validating the atom, I also added simple checks to
ensure that framerate's numerator and denominator as well as the
resolution are non-zero.