Add live composition to the compositor.
This patch makes use of Membrane timers to allow live operation of the
compositor.

A timer with interval equal to seconds / frames is started when the element starts operating.

Whenever a frame is composed naturally (when all vidoes have frames ready for composition) the timer is restarted.

When the timer ticks, a frame is forcibly rendered, using previous frames as fallback in cases where we don't have a frame ready.

To protect against stale frames clogging the front of the queue, whenever a frame is rendered stale frames are removed from the front of the queue.