Add a (rudimentary) test for live composition.
A new element, `BadConnectionEmulator`, was added to simulate delays and
packet loss.

The `ComposeMultipleInputs` pipeline was altered to allow introducing an
additional filter between the decoder and the compositor.

A new test was added (`packet_loss.ex`). It is completely analogous to
the h264 pipeline test, but uses the `BadConnectionEmulator` and sets
the compositor to live mode.

This test doesn't verify much unfortunately. Since some frames are
randomly dropped or delayed, there is no reference result to compare to.