Change the compositor API to use `VideoLayout`.
`VideoLayout` is a new version of `VideoProperties` with a more
ergonomic API for setting the display size of a video.

The display size is now fully independent of video resolution specified in
caps.

Also you can specify a name when adding a new pad to the compositor. Ina
subsequent patch, this name will be used to dynamically update video
layouts.