Make the compostor handle adding pads dynamically in offline mode.
Previously, when a pad was added in offline mode, the compositor could
enter a sort of deadlock when it wouldn't spit out any frames.

This was fixed by marking new videos as 'just added'. This flag gets
unset when a video has had it's frame used for composition at least
once. Until this happens, the video gets ignored if it's frames are 'too
new' for composing.