Fix the bug in `update_transformations`.
The bug was that `update_transformations` returned a `{:ok, :ok}` tuple
on success. This was caused by the return type on the rust side being
`Result<rustler::Atom, CompositorError>`. It seems that if rustler sees
a NIF function with a return type different than `Result<rustler::Atom,
rustler::Error>`, it will use a different elixir serializer, thus
causing `Ok(atoms:ok())` to be serialized as `{:ok, :ok}`. This patch
addresses this behaviour and also adds a test that checks
`update_transformation`'s return type.