Make the corner rounding shader use anti-aliasing and be branchless.
This patch removes all if statements from the corner rounding shader in
favor of arithmetic operations (which is optimal for shaders).

The shader also uses a very basic form of anti-aliasing now. Basically
it applies the `smoothstep` function to the alpha value of pixels on the
corners, which "fades them" gently instead of cutting them off sharply.