This patch removes the realtime mode and fixes a bug.
The bug was that the compositor used to not render any frames when
sending an end of stream message to a video. This meant that in cases
where all videos but one had frames for composing and then the one video
received end of stream, frames would never be rendered.

While fixing this problem, several serious bugs were discovered in the
'realtime mode' feature. We decided to remove it altogether since noone
uses it anyway.