Ms 422 design scene graph on the rust side (#98)
* Add a very simple and work in progress version of Rust serialization.

Most of this will change, but this is a big step forward :)

* Make encoding transformations resemble actual proposed encoding process.

* Make encoding modules more similar to the real thing.

There is still some work here.

* Rework the module structure.

This changes the location of all `RustlerFriendly` modules in the module
tree. Instead of being located in

`Membrane.VideoCompositor.Scene.RustlerFriendly.X`,

they're now located in modules

`Membrane.VideoCompositor.Scene.Object.X.RustlerFriendly`

Additionally, these modules are now defined in the same file their
parent module is in, which allows keeping all relevant parts in the same
file.

* Rework how we deal with resolution.

There is now a centralized function for converting a resolution to a
rustler-friendly tuple.

On the rust side Layout and Texture resolutions are now separate enums
to prevent putting a `TransformedInputResolution` into layout's
resolution.

* Change the RustlerFriendly structs to be better for rust translation.

* Implement scene parsing into a graph.

* Add missing elixir documentation.

* Add another rust test.

This one checks the transformation order after parsing the graph.

* Add documentation in rust.

I decided to keep the documentation for structs that exist on both sides
in the elixir code.

* Add old examples as end-to-end scene deserialization tests.

* Implement various small style fixes.

* Separate scene validation logic into a separate module

* Move transformation encoding logic to the `Transformation` module.

* Various style improvements again.

* Fix missing documentation for `Layout.rust_representation`

* Reorder some structs and modules for better readibility.

* Slight style improvements in some functions.

* Move validation functions into a new struct, `SceneValidator`.

This seems better than having two separate `impl Scene`s.

* Rename `contains_cycle` to `check_for_cycles`

* Fix the documentation of `Object.object_output_resolution`.

* Fix the documentation of `Layout.inputs`

* Implement the deserialization of `InputImage`s.

This includes inventing a system which would allow erasing the rustler
types from the implementation, so that it can be tested more rigorously.